fix: core/supabase/auth/servicesに移動し、コードのリファクタリングをしました
