+++ lib/core/purchase/services/revenue_cat_service.dart
@@ -1,7 +1,7 @@
 import 'dart:io';
 
 import 'package:flutter/services.dart';
-import 'package:food_gram_app/core/data/supabase/auth/account_service.dart';
+import 'package:food_gram_app/core/supabase/auth/services/account_service.dart';
 import 'package:food_gram_app/env.dart';
 import 'package:food_gram_app/main.dart';
 import 'package:purchases_flutter/purchases_flutter.dart';
