+++ lib/ui/screen/new_account/new_account_view_model.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:food_gram_app/core/data/supabase/auth/account_service.dart';
+import 'package:food_gram_app/core/supabase/auth/services/account_service.dart';
 import 'package:food_gram_app/core/utils/provider/loading.dart';
 import 'package:food_gram_app/ui/screen/new_account/new_account_state.dart';
 import 'package:riverpod_annotation/riverpod_annotation.dart';
